requirements.txt
# Consolidated Requirements for Rockfall Prediction System
# Production-ready dependencies optimized for performance

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database - MongoDB with Beanie ODM
motor==3.3.2
pymongo==4.6.0
beanie==1.23.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# Validation & Serialization
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# AI/ML Libraries
torch==2.1.0
torchvision==0.16.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.25.2
xgboost==2.0.2
lightgbm==4.0.0
shap==0.43.0
joblib==1.3.0

# Geospatial Processing
rasterio==1.3.9
geopandas==0.14.1
shapely==2.0.2
pyproj==3.6.1
folium==0.15.1

# Image Processing & Computer Vision
Pillow==10.1.0
opencv-python==4.8.1.78
matplotlib==3.8.2
seaborn==0.12.0
plotly==5.15.0

# Data Science & Analysis
scipy==1.10.0
tensorboard==2.13.0

# Export & Reporting
reportlab==4.0.7
openpyxl==3.1.2
fpdf2==2.7.6

# Communication & Notifications
twilio==8.11.0
sendgrid==6.10.0

# Background Processing
celery==5.3.4
redis==5.0.1

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Environment
python-dotenv==1.0.0